BUGFIX: Change C-S-Space keyboard mapping.
On Windows.
I forgot to convert no-break space to the line character set, this isn't
noticable on ISO8859-X character sets but sends a bad code for UTF-8
line set.  Generate a unicode no-break space for Ctrl-Shift-Space and
convert it to the line set rather than sending byte 160 directly up the
line to the server.