Collect DCS, APC, PM and SOS sequences.
This currently does nothing except eat the sequence.  The reason for it's
use is to allow us up increase the version number we declare in the DA
string without our screen getting messed up when someone assumes that
it's an XTerm version number.

The reason we want to increase the version number is to prompt vim to
allow 'sgr' mouse reporting.