Add Ansi wcwidth override sequence
ECMA-48: 8.3.90 PEC - PRESENTATION EXPAND OR CONTRACT

CSI Pn SP Z used for character width override.
    Only applies to characters that are not in US-ASCII.
    Does not work with DBCS fonts, only unicode conversion.

    This is a mode setting so applies to all following
    characters until reset.

    Pn=0   Default
    Pn=1   Force Wide for unicode chars.
    Pn=2   Force narrow for unicode chars.
    Pn=5   Force one Wide unicode char.
    Pn=6   Force one narrow unicode char.

This allows the host to do a DBCS to UTF-8 conversion and keep the
anachronistic alphabetic artifacts of the original encoding.

Note: The reset after one character for values 5&6 works by shifting the
value right by three bits if the bit-2 is set. This can also be used to
force one character narrow in a stream that normally wide or specifically
control the widths of a short precounted sequence of characters (upto 10).